+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - feat: add permissions to queue
   - feat: any user can get the list of users and groups
   - feat: logged in users can get the list of their groups
+  - feat: one cannot remove their own account
 
 ### Version 0.5.6 (2015-05-13)