+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ### Version 0.6 (2015-xx-xx)
 
   - feat: add permissions to queue
-  - feat: any user can get the list of users and groups
-  - feat: logged in users can get the list of their groups
-  - feat: one cannot remove their own account
+  - feat: regular users can get the list of users and groups
+  - feat: regular users can get the list of their groups
+  - feat: admin users can remove accounts (except their own)
 
 ### Version 0.5.6 (2015-05-13)