+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ### Version 0.6 (2015-xx-xx)
 
   - feat: add permissions to queue
+  - feat: non-destructive queue picking
   - feat: regular users can get the list of users and groups
   - feat: regular users can get the list of their groups
-  - feat: admin users can remove accounts (except their own)
+  - feat: admin users can remove any account but their own
   - feat: daily log rotation
+  - BREAKING: getting a queue no longer returns its elements (use new picking routes instead)
+  - BREAKING: only admin users can create queues
+  - BREAKING: queue push returns "success" message instead of queues 
 
 ### Version 0.5.6 (2015-05-13)